fetch only token in loader
