--- client/src/pages/Research.jsx
@@ -1,70 +1,0 @@
-import { useState, useEffect } from "react";
-import SearchBar from "../components/SearchBar";
-import "../App.css";
-
-const clientId = import.meta.env.VITE_CLIENT_ID;
-const clientSecret = import.meta.env.VITE_CLIENT_SECRET;
-
-function Research() {
-  const [accessToken, setAccessToken] = useState("");
-  const [albums, setAlbums] = useState([]);
-
-  useEffect(() => {
-    const authParameters = {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-      },
-      body: `grant_type=client_credentials&client_id=${clientId}&client_secret=${clientSecret}`,
-    };
-    fetch("https://accounts.spotify.com/api/token", authParameters)
-      .then((result) => result.json())
-      .then((data) => setAccessToken(data.access_token));
-  }, []);
-
-  async function searchAlbums(searchTerm) {
-    if (!accessToken) return;
-
-    const searchParameters = {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${accessToken}`,
-      },
-    };
-
-    const idResponse = await fetch(
-      `https://api.spotify.com/v1/search?q=${searchTerm}&type=artist`,
-      searchParameters
-    );
-    const artistData = await idResponse.json();
-    const artistID = artistData.artists.items[0].id;
-
-    const albumsResponse = await fetch(
-      `https://api.spotify.com/v1/artists/${artistID}/albums?include_groups=album&market=US&limit=50`,
-      searchParameters
-    );
-    const albumsData = await albumsResponse.json();
-    setAlbums(albumsData.items);
-  }
-
-  const searchResult = searchAlbums;
-  return (
-    <>
-      <SearchBar searchResult={searchResult} />
-
-      {albums.map((album) => (
-        <section className="artist-card" key={album.id}>
-          <img
-            className="artist-picture"
-            src={album.images[0]?.url}
-            alt={album.name}
-          />
-          <h3 className="albumName">{album.name}</h3>
-        </section>
-      ))}
-    </>
-  );
-}
-
-export default Research;
